dogrunのIDをint64に変更(#130)

+++ internal/dogrun/core/dto/dogrun_res_dto.go
@@ -6,8 +6,8 @@ import (
 
 // ドッグラン詳細画面での表示情報
 type DogrunDetail struct {
-	DogrunID        int          `json:"dogrunId,omitempty"`
-	DogrunManagerID int          `json:"dogrunManagerId,omitempty"`
+	DogrunID        int64        `json:"dogrunId,omitempty"`
+	DogrunManagerID int64        `json:"dogrunManagerId,omitempty"`
 	PlaceId         string       `json:"placeId,omitempty"`
 	Name            string       `json:"name"`
 	Address         Address      `json:"address"`
@@ -25,7 +25,7 @@ type DogrunDetail struct {
 
 // ドッグラン一覧での表示情報
 type DogrunLists struct {
-	DogrunID          int             `json:"dogrunId,omitempty"`
+	DogrunID          int64           `json:"dogrunId,omitempty"`
 	PlaceId           string          `json:"placeId,omitempty"`
 	Name              string          `json:"name"`
 	Address           Address         `json:"address"`
